readjusted folder structure of utils
