--- src/utils/sort.ts
@@ -1,70 +1,0 @@
-import { BlogCard } from "./data/posts/CardData";
-import { BlogExcerpt } from "./data/posts/ExcerptData";
-import { GithubRepo } from "./data/repos/GithubRepoData";
-
-type SortFormat = "newest" | "oldest";
-
-type AcceptableContent = BlogCard[] | BlogExcerpt[];
-
-export function dateSort(format: SortFormat, content: BlogCard[]): BlogCard[];
-export function dateSort(format: SortFormat, content: BlogExcerpt[]): BlogExcerpt[];
-export function dateSort(format: SortFormat, content: AcceptableContent) {
-  if (format === "newest") {
-    return content.sort((a, b) => {
-      const aDate = new Date(a.publishDate);
-      const bDate = new Date(b.publishDate);
-
-      if (aDate > bDate) {
-        return -1;
-      } else if (aDate < bDate) {
-        return 1;
-      } else {
-        return 0;
-      }
-    });
-  } else if (format === "oldest") {
-    return content.sort((a, b) => {
-      const aDate = new Date(a.publishDate);
-      const bDate = new Date(b.publishDate);
-
-      if (aDate > bDate) {
-        return 1;
-      } else if (aDate < bDate) {
-        return -1;
-      } else {
-        return 0;
-      }
-    });
-  }
-  throw new Error("Incorrect sort format");
-}
-
-export function repoSort(format: SortFormat, content: GithubRepo[]) {
-  if (format === "newest") {
-    return content.sort((a, b) => {
-      const aDate = new Date(a.createdAt);
-      const bDate = new Date(b.createdAt);
-
-      if (aDate > bDate) {
-        return -1;
-      } else if (aDate < bDate) {
-        return 1;
-      } else {
-        return 0;
-      }
-    });
-  } else {
-    return content.sort((a, b) => {
-      const aDate = new Date(a.createdAt);
-      const bDate = new Date(b.createdAt);
-
-      if (aDate > bDate) {
-        return 1;
-      } else if (aDate < bDate) {
-        return -1;
-      } else {
-        return 0;
-      }
-    });
-  }
-}
